uses current transformers and updates them to follow structure. Deletes activity transformer
