+++ app/Http/Transformers/EventTransformer.php
@@ -15,10 +15,18 @@ class EventTransformer extends TransformerAbstract
      */
     public function transform(Event $event)
     {
-        return [
-            'event_id' => (string) $event->id,
+         return [
+            'event_id' => $event->id,
             'event_type' => $event->event_type,
             'submission_type' => $event->submission_type,
+            'quantity' => $event->quantity,
+            'quantity_pending' => $event->quantity_pending,
+            'why_participated' => $event->why_participated,
+            'caption' => $event->caption,
+            'status' => $event->status,
+            'source' => $event->source,
+            'remote_addr' => $event->remote_addr,
+            'reason' => $event->reason,
             'created_at' => $event->created_at->toIso8601String(),
             'updated_at' => $event->updated_at->toIso8601String(),
         ];
