+++ app/Http/Transformers/SignupTransformer.php
@@ -3,6 +3,7 @@
 namespace Rogue\Http\Transformers;
 
 use Rogue\Models\Signup;
+use Rogue\Models\Event;
 use League\Fractal\TransformerAbstract;
 
 class SignupTransformer extends TransformerAbstract
@@ -13,8 +14,8 @@ class SignupTransformer extends TransformerAbstract
      * @var array
      */
     protected $defaultIncludes = [
-        'post',
-        'event',
+        'posts',
+        'events',
     ];
 
     /**
@@ -26,29 +27,19 @@ class SignupTransformer extends TransformerAbstract
     public function transform(Signup $signup)
     {
         return [
-            'id' => $signup->id,
+            'signup_id' => $signup->id,
+            'signup_event_id' => $signup->event_id,
+            'submission_type' => $signup->event->submission_type,
             'northstar_id' => $signup->northstar_id,
             'campaign_id' => $signup->campaign_id,
             'campaign_run_id' => $signup->campaign_run_id,
             'quantity' => $signup->quantity,
             'quantity_pending' => $signup->quantity_pending,
             'why_participated' => $signup->why_participated,
+            'signup_source' => $signup->source,
             'created_at' => $signup->created_at->toIso8601String(),
             'updated_at' => $signup->updated_at->toIso8601String(),
         ];
-    }
-
-    /**
-     * Include the event
-     *
-     * @param \Rogue\Models\Signup $signup
-     * @return \League\Fractal\Resource\Item
-     */
-    public function includeEvent(Signup $signup)
-    {
-        $event = $signup->event;
-
-        return $this->item($event, new EventTransformer);
     }
 
     /**
@@ -57,10 +48,25 @@ class SignupTransformer extends TransformerAbstract
      * @param \Rogue\Models\Signup $signup
      * @return \League\Fractal\Resource\Collection
      */
-    public function includePost(Signup $signup)
+    public function includePosts(Signup $signup)
     {
         $post = $signup->posts;
 
         return $this->collection($post, new PostTransformer);
     }
+
+    /**
+     * Include the event
+     *
+     * @param \Rogue\Models\Signup $signup
+     * @return \League\Fractal\Resource\Item
+     */
+    public function includeEvents(Signup $signup)
+    {
+        // $event = $signup->event;
+        $event = Event::where('signup_id', $signup->id)->get();
+
+        return $this->collection($event, new EventTransformer);
+    }
+
 }
